lesson-10-observability/demo-2/requirements.txt
google-adk>=1.17.0
google-search-results